+++ lib/hexapdf/type/acro_form/variable_text_field.rb
@@ -34,20 +34,31 @@
 # commercial licenses are available at <https://gettalong.at/hexapdf/>.
 #++
 
+require 'hexapdf/dictionary'
+require 'hexapdf/stream'
+
 module HexaPDF
   module Type
-
-    # Namespace module for all AcroForm related dictionary types.
-    #
-    # See: PDF1.7 s12.7
     module AcroForm
 
-      autoload(:Form, 'hexapdf/type/acro_form/form')
-      autoload(:Field, 'hexapdf/type/acro_form/field')
-      autoload(:VariableTextField, 'hexapdf/type/acro_form/variable_text_field')
-      autoload(:TextField, 'hexapdf/type/acro_form/text_field')
+      # An AcroForm variable text field defines how text that it is not known at generation time
+      # should be rendered. For example, AcroForm text fields (normally) don't have an initial
+      # value; the value is entered by the user and needs to be rendered correctly by the PDF
+      # reader.
+      #
+      # See: PDF1.7 s12.7.3.3
+      class VariableTextField < Field
+
+        define_field :DA, type: String
+        define_field :Q, type: Integer, default: 0, allowed_values: [0, 1, 2]
+        define_field :DS, type: String, version: '1.5'
+        define_field :RV, type: [String, Stream], version: '1.5'
+
+        # All inheritable dictionary fields for text fields.
+        INHERITABLE_FIELDS = (superclass::INHERITABLE_FIELDS + [:DA, :Q]).freeze
+
+      end
 
     end
-
   end
 end