+++ test/hexapdf/type/acro_form/test_text_field.rb
@@ -1,0 +1,22 @@
+# -*- encoding: utf-8 -*-
+
+require 'test_helper'
+require 'hexapdf/document'
+require 'hexapdf/type/acro_form/text_field'
+
+describe HexaPDF::Type::AcroForm::TextField do
+  before do
+    @doc = HexaPDF::Document.new
+    @field = @doc.add({}, type: :XXAcroFormField, subtype: :Tx)
+  end
+
+  it "resolves /MaxLen as inheritable field" do
+    assert_nil(@field[:MaxLen])
+
+    @field[:Parent] = {MaxLen: 5}
+    assert_equal(5, @field[:MaxLen])
+
+    @field[:MaxLen] = 6
+    assert_equal(6, @field[:MaxLen])
+  end
+end