+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Convenience methods for accessing field flags for
   [HexaPDF::Type::AcroForm::Field]
+* [HexaPDF::Type::AcroForm::TextField] and
+  [HexaPDF::Type::AcroForm::VariableTextField] for basic text field support
 
 
 ## 0.11.9 - 2020-06-15